Change uap10.0.16299 to uap10.0.18362
Visual Studio 2022 doesn't support Windows SDK 10.0.16299 - the oldest
it supports is 10.0.18362. The last version of Windows for which the
latest available surface area was 16299 (Windows 10 1709) went out of
support in April 2019. SDK 18362 started with Windows 10 1903 in May
2019 (and 1903 itself went out of support in November 2020), so this is
still a fairly old API version, accommodating some long-out-of-date
versions of Windows. This was the smallest increment that would bring us
to SDK versions supported by the current tooling.