Replace xUnit with mstest (#1883)
xUnit is no longer able to run tests on UWP. MSTest supports this in Visual Studio 2022, so we are moving back to MSTest. (Rx used to used MSTest before, so this was a relatively straightforward change.)

We had to make a few changes due to the fact that xUnit supplies a SynchronizationContext whereas MSTest does not.

Also fixed a race condition in CreateAsyncTest.