Exploit unconstrained type parameter annotations
This enables us to re-enable warnings CS8600, CS8602, CS8603, CS8604, and CS8632.

When support for nullable reference types was added to Rx in 2020, you were not allowed to write `T?` if `T` was an unconstrained type parameter. Although that made sense - if `T` could be either a value type or a reference type, `T?` could mean fundamentally different things for different type arguments - in practice it was unhelpful because it meant there was no straightforward way to indicate "nullable in cases where T is a reference type", which is often what was required in practice.

C# 9.0 added a feature that removed this limitation:
https://github.com/dotnet/csharplang/blob/main/proposals/csharp-9.0/unconstrained-type-parameter-annotations.md

This commit takes advantage of this language feature to better express some of the existing null handling.