Add large mode to CompositeDisposable
Once the number of elements in a CompositeDisposable gets into the thousands, Remove became very expensive because it performed a linear search. This change causes CompositeDisposable to switch into an alternate mode using a dictionary to enable hash-based lookup once it grows over a particular size.

With under 1024 elements, the behaviour is as before. The number of fields remains the same so this should not change the memory footprint of small CompositeDisposables. However, it does require it to check which of the two modes it is in, so there will likely be a small performance penalty for small CompositeDisposables, so we will need to see if this changes any benchmarks in a significant way before merging this change.