Resolve SDK 8.0.2xx diagnostics (#2100)
It looks like the 8.0.200 .NET SDK added a lot of new diagnostics suggesting the use of collection expressions.

For the most part we have taken these suggestions, but in some cases we've retained the existing code so you can see what types are being used. I find that in methods with lots of overloads (e.g., Assert.AreEqual) it becomes quite difficult to work out what will actually happen if you replace an explicitly typed list initializer with just `[...]`.