Enable application-defined Synchronize gate
Our AsyncGate does not support cancellation, and for now at least we don't want to add it. (For one thing, it opens the can of worms of whether we want to attempt to support cancellation across the board in AsyncRx.NET. But also, more or less everyone who tries to add cancellation support to this sort of primitive ends up creating subtle bugs.)

So this defines an IAsyncGate interface and Synchronize overloads that accept it, enabling them to work with application-supplied implementations.