Fix bug where short-time Schedule ran synchronously
There were situations in which Scheduling a timed work item in the TaskPoolScheduler would occasionally end up running the work before returning from Schedule. This was not intentional, and occurred due small window in which a Delay task could complete before a continuation was attached to it. (This only seems to have started happening in .NET 9.0.)

#2207
Also I realised there were bugs in the relevant tests: they had Asserts inside the work item callbacks, but those assertions would fire on threads not known to the unit test framework, and nothing in the test code attempted to await the work and rethrow back on the test thread. So I restructured these to move all of the assertions back into the main test bodies.