Update xunit.assert.source to 2.8.1 and remove workaround
We had a workaround for https://github.com/xunit/xunit/issues/2703 but this has been fixed, so by upgrading we can remove the workaround.

This is good because it turns out the workaround doesn't work if you have a machine with a dev drive user via an alias folder. (Some parts of the build system resolve to the unaliased dev drive path, and so our <Compile Remove> didn't actually match the thing it's meant to be removing!)