Finish unobsoleting most of the UI support types
Since System.Reactive now has to become a legacy facade, we don't need to go through the contortions required to gradually deprecate the UI types it contains. The new UI-specific packages now contain these types and System.Reactive is now just a facade for these. (With a couple of exceptions - UWP's 'special' ThreadPoolScheduler is problematic, and RemotingObservable is one we're deliberately relegating to only being available through the legacy facade to find out whether anyone is really using it, because nobody should be doing .NET Remoting any more. If necessary we'll introduce a System.Reactive.For.Remoting but I'd prefer not to.)