Fix ref assembly build
Turns out the solution doesn't build the ref assemblies and they weren't building.

I also needed to modify the async code generator to correctly handle #if-ed out code. Turns out it was stripping the directives and then copying the disabled block of text in as leading trivia! This had the effect of copying #if-ed out methods into the generated code, making them reappear.

I've also worked through the first four files of operators.