Fix discrepancies in Obsolete attributes
In some cases, the new System.Linq.AsyncEnumerable simply doesn't supply equivalents to the projection-based overloads we've always offered. And in some of these cases we were directing people to recreate the functionality with Select, when in fact we continue to offer equivalent functionality in System.Interactive.Async

Also reinstated some operations that got left out.

#2248